🎨[style]: 파일 정보 주석 변경 및 네이밍 변경
