+++ OpenMarket/OpenMarket/Controller/RegistrationViewController.swift
@@ -1,8 +1,8 @@
 //
-//  ProductRegistrationViewController.swift
+//  RegistrationViewController.swift
 //  OpenMarket
 //
-//  Created by song on 2022/05/25.
+//  Created by marlang, Taeangel on 2022/05/25.
 //
 
 import UIKit
