+++ OpenMarket/OpenMarket/Controller/enum/CurrencyType.swift
@@ -2,7 +2,7 @@
 //  CurrencyType.swift
 //  OpenMarket
 //
-//  Created by 김동욱 on 2022/05/27.
+//  Created by marlang, Taeangel on 2022/05/27.
 //
 
 import Foundation
