+++ OpenMarket/OpenMarket/Controller/enum/LayoutType.swift
@@ -2,7 +2,7 @@
 //  LayoutType.swift
 //  OpenMarket
 //
-//  Created by 김동욱 on 2022/05/19.
+//  Created by marlang, Taeangel on 2022/05/19.
 //
 
 import UIKit
