+++ OpenMarket/OpenMarket/Utils/Extension/Data.swift
@@ -2,7 +2,7 @@
 //  Data.swift
 //  OpenMarket
 //
-//  Created by song on 2022/05/25.
+//  Created by marlang, Taeangel on 2022/05/25.
 //
 
 import UIKit
