+++ OpenMarket/OpenMarket/Utils/Extension/String.swift
@@ -2,7 +2,7 @@
 //  String.swift
 //  OpenMarket
 //
-//  Created by 김동욱 on 2022/05/18.
+//  Created by marlang, Taeangel on 2022/05/18.
 //
 
 import UIKit
