+++ OpenMarket/OpenMarket/Utils/Extension/UIImageView.swift
@@ -2,15 +2,10 @@
 //  UIImageView.swift
 //  OpenMarket
 //
-//  Created by 김동욱 on 2022/05/20.
+//  Created by marlang, Taeangel on 2022/05/20.
 //
 
 import UIKit
-
-struct Cache {
-    static let cache = NSCache<NSURL, UIImage>()
-    private init() {}
-}
 
 extension UIImageView {
