+++ OpenMarket/OpenMarket/Utils/Extension/UIStackView.swift
@@ -2,7 +2,7 @@
 //  UIStackView.swift
 //  OpenMarket
 //
-//  Created by song on 2022/05/17.
+//  Created by marlang, Taeangel on 2022/05/17.
 //
 
 import UIKit
