+++ OpenMarket/OpenMarket/Utils/Singletone/Cache.swift
@@ -1,0 +1,13 @@
+//
+//  Cache.swift
+//  OpenMarket
+//
+//  Created by marlang, Taeangel on 2022/06/01.
+//
+
+import UIKit
+
+struct Cache {
+    static let cache = NSCache<NSURL, UIImage>()
+    private init() {}
+}
