+++ OpenMarket/OpenMarket/View/Cell/CustomCell.swift
@@ -2,7 +2,7 @@
 //  CustomCell.swift
 //  OpenMarket
 //
-//  Created by 김동욱 on 2022/05/20.
+//  Created by marlang, Taeangel on 2022/05/20.
 //
 
 import UIKit
