+++ OpenMarket/OpenMarket/View/Cell/GridCell.swift
@@ -2,7 +2,7 @@
 //  GridCell.swift
 //  OpenMarket
 //
-//  Created by song on 2022/05/17.
+//  Created by marlang, Taeangel on 2022/05/17.
 //
 
 import UIKit
