+++ OpenMarket/OpenMarket/View/Cell/ListCell.swift
@@ -2,7 +2,7 @@
 //  ListCell.swift
 //  OpenMarket
 //
-//  Created by song on 2022/05/17.
+//  Created by marlang, Taeangel on 2022/05/17.
 //
 
 import UIKit
